[Reaf] EnhancedInput Function(Character Action) [Controller -> PlayerCharacter]
