+++ Source/ProejctUN/Character/UNPlayerCharacter.cpp
@@ -2,4 +2,102 @@
 
 
 #include "../Character/UNPlayerCharacter.h"
+#include "AbilitySystemComponent.h"
+#include "../Player/UNGASPlayerState.h"
+#include "EnhancedInputComponent.h"
+#include "InputMappingContext.h"
+#include "InputActionValue.h"
+#include "EnhancedInputSubsystems.h"
+#include "Blueprint/AIBlueprintHelperLibrary.h"
+
+AUNPlayerCharacter::AUNPlayerCharacter()
+{
+	ASC = nullptr;
+
+	CachedDestination = FVector::ZeroVector;
+	FollowTime = 0.f;
+	ShortPressThreshold = 0.3f;
+
+	static ConstructorHelpers::FObjectFinder<UInputMappingContext> InputMappingContextRef(TEXT("/Script/EnhancedInput.InputMappingContext'/Game/Input/IMC_Character.IMC_Character'"));
+	if (nullptr != InputMappingContextRef.Object)
+	{
+		DefaultMappingContext = InputMappingContextRef.Object;
+	}
+
+	static ConstructorHelpers::FObjectFinder<UInputAction> InputActionMoveRef(TEXT("/Script/EnhancedInput.InputAction'/Game/Input/Action/IA_Move.IA_Move'"));
+	if (nullptr != InputActionMoveRef.Object)
+	{
+		SetDestinationClickAction = InputActionMoveRef.Object;
+	}
+}
+
+UAbilitySystemComponent* AUNPlayerCharacter::GetAbilitySystemComponent() const
+{
+	return ASC;
+}
+
+void AUNPlayerCharacter::SetupPlayerInputComponent(UInputComponent* PlayerInputComponent)
+{
+	Super::SetupPlayerInputComponent(PlayerInputComponent);
+
+	if (UEnhancedInputComponent* EnhancedInputComponent = Cast<UEnhancedInputComponent>(InputComponent))
+	{
+		EnhancedInputComponent->BindAction(SetDestinationClickAction, ETriggerEvent::Started, this, &AUNPlayerCharacter::OnInputStarted);
+		EnhancedInputComponent->BindAction(SetDestinationClickAction, ETriggerEvent::Triggered, this, &AUNPlayerCharacter::OnSetDestinationTriggered);
+		EnhancedInputComponent->BindAction(SetDestinationClickAction, ETriggerEvent::Completed, this, &AUNPlayerCharacter::OnSetDestinationReleased);
+		EnhancedInputComponent->BindAction(SetDestinationClickAction, ETriggerEvent::Canceled, this, &AUNPlayerCharacter::OnSetDestinationReleased);
+	}
+}
+
+void AUNPlayerCharacter::BeginPlay()
+{
+	Super::BeginPlay();
+
+	APlayerController* PlayerController = CastChecked<APlayerController>(GetController());
+	if (UEnhancedInputLocalPlayerSubsystem* Subsystem = ULocalPlayer::GetSubsystem<UEnhancedInputLocalPlayerSubsystem>(PlayerController->GetLocalPlayer()))
+	{
+		Subsystem->AddMappingContext(DefaultMappingContext, 0);
+	}
+}
+
+void AUNPlayerCharacter::OnInputStarted()
+{
+	APlayerController* PlayerController = CastChecked<APlayerController>(GetController());
+	PlayerController->StopMovement();
+}
+
+void AUNPlayerCharacter::OnSetDestinationTriggered()
+{
+	APlayerController* PlayerController = CastChecked<APlayerController>(GetController());
+	FollowTime += GetWorld()->GetDeltaSeconds();
+
+	FHitResult Hit;
+	bool bHitSuccessful = false;
+	bHitSuccessful = PlayerController->GetHitResultUnderCursor(ECollisionChannel::ECC_Visibility, true, Hit);
+
+	if (bHitSuccessful)
+	{
+		CachedDestination = Hit.Location;
+		GEngine->AddOnScreenDebugMessage(-1, 1.f, FColor::Green, FString::Printf(TEXT("%f"), CachedDestination.X));
+	}
+
+
+	FVector WorldDirection = (CachedDestination - GetActorLocation()).GetSafeNormal();
+	AddMovementInput(WorldDirection, 1.0, false);
+	GEngine->AddOnScreenDebugMessage(-1, 1.f, FColor::Green, FString::Printf(TEXT("%f"), FollowTime));
+
+}
+
+void AUNPlayerCharacter::OnSetDestinationReleased()
+{
+	if (FollowTime <= ShortPressThreshold)
+	{
+		APlayerController* PlayerController = CastChecked<APlayerController>(GetController());
+		UAIBlueprintHelperLibrary::SimpleMoveToLocation(PlayerController, CachedDestination);
+	}
+
+	FollowTime = 0.f;
+
+	GEngine->AddOnScreenDebugMessage(-1, 1.f, FColor::Green, FString::Printf(TEXT("Released!")));
+}
